Add H1 tag
The trix core team explains how you can add an H1 option at runtime.

https://github.com/basecamp/trix/issues/133

This worked for initial editing, but when loading previously saved markup
with H1 tags, the H1 tags would be removed.

It is faster make these simple additions to the trix source than to
trying to debug the loading issues.